Separate models into file
