+++ lib/middleman-blog-similar/models/tagging.rb
@@ -1,0 +1,10 @@
+module Middleman
+  module Blog
+    module Similar
+      class Tagging < ActiveRecord::Base
+        belongs_to :article
+        belongs_to :tag
+      end
+    end
+  end
+end
